Transfer primer class tests
